Move LocalPingableLeader out of PaxosLeaderElectionServiceBuilder
